backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
pydantic
python-multipart
PyPDF2
openai
python-dotenv
nltk
pdfminer.six
wordfreq